// sdk/packages/core/src/nexus-core/types.ts
export interface NexusCoreStartInput
	extends Omit<StartSessionInput, "config" | "localRuntime"> {
	config: NexusCoreStartConfig;
	localRuntime?: LocalRuntimeStartOptions;
}